Improve the performance of the rect fill command
