+++ Source/Libs/Gfx/Gfx_DrawLine.c
@@ -26,36 +26,45 @@
 	
 */
 
-#include "../Libs/Include/Heap.h"
-#include "../Libs/Include/Bitmap.h"
-#include "../Libs/Include/Gfx.h"
 
-extern U16 _Kernel_LowMemoryInfoKiB;
+#include "../Include/Gfx.h"
 
 
-static HeapArea* _Kernel_Heap;
-
-
-static void _InitializeHeap() {
-  void* heapStart = (void*)0xf000;
-  U32 heapSize = (_Kernel_LowMemoryInfoKiB  - 64) * 1024;
-  U16 blockSize = 8;
-
-  _Kernel_Heap = Heap.Initialize(heapStart, heapSize, blockSize);
-}
-
-void KernelMain() {
-  _InitializeHeap();
-
-  if (Gfx.Core.Initialize(640, 480, 4, _Kernel_Heap))
-    Gfx.Core.RefreshFromBackBuffer();
-  
-  Gfx.Draw.FilledRect(5, 5, 630, 470, 3);
-  Gfx.Draw.String(10, 10, "Hello, World!", 15);
-  Gfx.Core.RefreshFromBackBuffer();
+// Draws a line using Bresenham's algorithm
+__attribute__((unused))
+void Gfx_DrawLine(U16 startX, U16 startY, U16 endX, U16 endY, U8 colorIndex) {
+  I16 dx = (endX > startX)
+    ? (endX - startX)
+    : (startX - endX);
     
- 
-  while (1)
-    __asm__ __volatile__("hlt");
+  I16 sx = (startX < endX)
+    ? 1
+    : -1;
+    
+  I16 dy = (endY > startY)
+    ? (startY - endY)
+    : (endY - startY);
+    
+  I16 sy = (startY < endY)
+    ? 1
+    : -1;
+    
+  I16 err = dx + dy;
+    
+  while (1) {
+    Gfx.Draw.Pixel(startX, startY, colorIndex);
+    if (startX == endX && startY == endY)
+      break;
+	
+    I16 e2 = 2 * err;
+    if (e2 >= dy){
+      err += dy;
+      startX += sx;
+    }
+    
+    if (e2 <= dx) {
+      err += dx;
+      startY += sy;
+    }
+  }
 }
-
